# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi documentación de Minibares",
        "description": "Class LogController",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tipos_posteos": {
            "get": {
                "tags": [
                    "TiposPosteos"
                ],
                "summary": "List all tipos_posteos",
                "description": "Display a listing of the resource.",
                "operationId": "b9ebea6e77dcf808b8b8c360de417f8b",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "TiposPosteos"
                ],
                "summary": "Crea un tipo de posteo",
                "description": "Metodo para crear un tipo de posteo",
                "operationId": "00f99d6b4afb0d67a7ed424ac0d713b8",
                "requestBody": {
                    "description": "Objeto JSON que contiene los datos para crear un nuevo tipo de posteo",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Tipo de posteo creado con exito."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/tipos_posteos/{id}": {
            "get": {
                "tags": [
                    "TiposPosteos"
                ],
                "summary": "Obtiene un tipo de posteo",
                "description": "Metodo para obtener un tipo de posteo.",
                "operationId": "162cd7b39a560225064bf9aa38d92857",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del tipo",
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "description": "Descripcion del tipo",
                                        "type": "string"
                                    },
                                    "facturaSN": {
                                        "description": "Indica si es un valor a añadir en la factura o no",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de posteo obtenido con exito."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "TiposPosteos"
                ],
                "summary": "Actualiza un tipo de posteo",
                "description": "Metodo para actualizar un tipo de posteo",
                "operationId": "578fd642834b964182a27ce3af659d27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del tipo de posteo.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del tipo",
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "description": "Descripcion del tipo",
                                        "type": "string"
                                    },
                                    "facturaSN": {
                                        "description": "Indica si es un valor a añadir en la factura o no",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de posteo actualizado con exito."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api//tipos_posteos/{id}": {
            "delete": {
                "tags": [
                    "TiposPosteos"
                ],
                "summary": "Elimina un tipo de posteo",
                "description": "Metodo para eliminar un tipo de posteo",
                "operationId": "1ada80c2a1cf64b20f3e3540da9d7a9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del tipo de posteo.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de posteo eliminado con exito."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    }
}